+++ src/register-commands-options.ts
@@ -1,0 +1,7 @@
+import { Command } from './command';
+
+export interface RegisterCommandsOptions {
+	commands: (typeof Command)[];
+	name?: string;
+	forceExit?: boolean;
+}