+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 ## [1.1.0]
 
 ### Additions
+- [Issue #1] Add forceExit option
 
 ### Fixes